+++ EffectFramework.Core/Models/Db/Tenant.cs
@@ -2,7 +2,7 @@
 {
     public class Tenant
     {
-        public int TenantID { get; set; }
+        public long TenantID { get; set; }
         public string Name { get; set; }
         public bool IsDeleted { get; set; }
     }
